🔨 Refactor(#28): partForm수정
