+++ src/features/part/apis.ts
@@ -2,15 +2,12 @@ import api from '../../apis/axios/instance';
 import { Part } from './types';
 
 const partsApis = {
-  getParts: async (): Promise<Part[]> => {
+  get: async (): Promise<Part[]> => {
     const response = await api.get('/parts');
     return response.data;
   },
-  postParts: async (part: Part): Promise<void> => api.post('/parts', part),
-  putPart: async (part: Part): Promise<void> => {
-    const { id, ...rest } = part;
-    return api.put(`/parts/${id}`, rest);
-  },
-  deletePart: async (id: number): Promise<void> => api.delete(`/parts/${id}`),
+  post: async (params: Omit<Part, 'id'>): Promise<void> =>
+    api.post('/parts', params),
+  delete: async (id: number): Promise<void> => api.delete(`/parts/${id}`),
 };
 export default partsApis;
