+++ src/pages/Part.tsx
@@ -8,23 +8,18 @@ import {
 } from 'react-bootstrap';
 import useModal from '../hooks/useModal';
 import PartForm from '../features/part/ui/PartForm';
-import { useState } from 'react';
-import usePartStore from '../store/usePartStore';
 import type { Part } from '../features/part/types';
 import partsQueries from '../features/part/queries';
 
 export default function Part() {
   const { isShow, openModal, closeModal, Modal } = useModal();
-  const [mod, setMod] = useState<'create' | 'update'>();
   const { data: parts } = partsQueries.read();
-  const createMutation = partsQueries.create();
-  const updateMutation = partsQueries.update();
   const deleteMutation = partsQueries.delete();
-  const { part, resetPart, setPart } = usePartStore();
+
   return (
     <>
       <Modal title="파트 추가" isShow={isShow} closeModal={closeModal}>
-        <PartForm />
+        <PartForm closeModal={closeModal} />
       </Modal>
 
       <Container>
@@ -33,7 +28,6 @@ export default function Part() {
             <Button
               type="button"
               onClick={() => {
-                setMod('create');
                 openModal();
               }}
             >
@@ -61,22 +55,8 @@ export default function Part() {
                       <Button
                         variant="secondary"
                         onClick={() => {
-                          setMod('update');
-                          setPart({
-                            id: part.id,
-                            sectionId: part.sectionId,
-                            name: part.name,
-                          });
-                          openModal();
-                        }}
-                      >
-                        수정
-                      </Button>
-                      <Button
-                        variant="secondary"
-                        onClick={() => {
-                          deleteMutation.mutate(part.id);
                           if (confirm('삭제 하시겠습니까?')) {
+                            deleteMutation.mutate(part.id);
                           }
                         }}
                       >
